+++ src/js/index.js
@@ -1,336 +1,20 @@
-/**
- * JS doc 타입 정의
- */
-
-/**
- * @typedef {"espresso" | "frappuccino" | "blended" | "teavana" | "dessert"} MenuCategory
- */
-
-/**
- * @typedef {Object} MenuObject
- * @property {String} name
- * @property {boolean} isSoldOut
- */
-
-/**
- * @typedef {Object} UseMenuReturn
- * @property {Object} MenuState
- * @property {Function} AddMenu
- */
-
-/**
- * @typedef {Object} AppendMenuOnClickCallbackObject
- * @property {Function} onClickSoldOut
- * @property {Function} onClickUpdate
- * @property {Function} onClickRemove
- */
-
-/**
- * Global Constants
- */
-
-const elementIdMap = {
-  espressoMenuForm: "espresso-menu-form",
-  espressoMenuList: "espresso-menu-list",
-  espressoMenuNameInput: "espresso-menu-name",
-  removeButton: "removeButton",
-  updateButton: "updateButton",
-  soldOutButton: "soldOutButton",
-  menuName: "menuName",
-  menuWrapper: "menuWrapper",
-  menuCategoryButtonWrapper: "menuCategoryButtonWrapper",
-  menuTitleName: "menuTitleName",
-};
-
-const localStorageKey = {
-  categoryState: "moonbucksState.categoryState",
-  categorySeqState: "moonbucksState.categorySeqState",
-};
-
-const localCategoryState = JSON.parse(
-  localStorage.getItem(localStorageKey.categoryState)
-);
-
-const localCategorySeqState = JSON.parse(
-  localStorage.getItem(localStorageKey.categorySeqState)
-);
-
-const saveInLocalStorageHandler = (localStorageKey) => ({
-  set(target, prop, val, receiver) {
-    Reflect.set(target, prop, val, receiver);
-    localStorage.setItem(localStorageKey, JSON.stringify(target));
-    return true;
-  },
-});
-
-const categoryState = new Proxy(
-  localCategoryState || {
-    espresso: {},
-    frappuccino: {},
-    blended: {},
-    teavana: {},
-    dessert: {},
-  },
-  saveInLocalStorageHandler(localStorageKey.categoryState)
-);
-
-const categorySeqState = new Proxy(
-  localCategorySeqState || {
-    espresso: 0,
-    frappuccino: 0,
-    blended: 0,
-    teavana: 0,
-    dessert: 0,
-  },
-  saveInLocalStorageHandler(localStorageKey.categorySeqState)
-);
+import useMenu from "./useMenu.js";
+import { elementIdMap } from "./utils/constant_utils.js";
+import { getById } from "./utils/control_dom_utils.js";
 
 const [getState, setState, incrementId, setCategoryName] = useMenu();
 
-/**
- * 메뉴의 상태를 만들고 관리한다.
- * @param {MenuCategory} initCategoryName
- * @returns {UseMenuReturn} it returns menu state and function to add menu.
- */
-function useMenu(initCategoryName = "espresso") {
-  /**
-   * Init
-   */
-  let categoryNameState;
-  let menuState;
-
-  setCategoryName(initCategoryName);
-
-  /**
-   * Mutation Functions
-   */
-
-  function setCategoryName(categoryName) {
-    if (menuState) {
-      stashMenu();
-    }
-
-    categoryNameState = categoryName;
-    setMenuState(categoryState[categoryName]);
-  }
-
-  function setMenuState(param) {
-    menuState = param;
-    categoryState[categoryNameState] = menuState;
-
-    renderMenu();
-  }
-
-  function remove(seq) {
-    if (confirm("정말로 삭제하시겠습니까?")) {
-      const { [seq]: removeMenu, ...rest } = menuState;
-
-      setMenuState({
-        ...rest,
-      });
-    }
-  }
-
-  function soldOut(seq) {
-    const { [seq]: soldOutMenu, ...rest } = menuState;
-    soldOutMenu.isSoldOut = !soldOutMenu.isSoldOut;
-
-    setMenuState({
-      [seq]: soldOutMenu,
-      ...rest,
-    });
-  }
-
-  function update(seq) {
-    const newName = prompt("수정하고 싶은 이름을 입력해주세요.");
-
-    if (newName) {
-      const { [seq]: updateMenu, ...rest } = menuState;
-      updateMenu.name = newName;
-
-      setMenuState({
-        [seq]: updateMenu,
-        ...rest,
-      });
-    }
-  }
-
-  /**
-   * Get Functions
-   */
-
-  function getMenuSeq() {
-    return ++categorySeqState[categoryNameState];
-  }
-
-  function getMenuState() {
-    return categoryState[categoryNameState];
-  }
-
-  function getMenuWrapperId(seq) {
-    const { menuWrapper } = elementIdMap;
-    return `${menuWrapper}-${seq}`;
-  }
-
-  /**
-   * UI Control Functions
-   */
-
-  function renderMenu() {
-    renderAppendMenu(menuState);
-    renderSoldOutMenu(menuState);
-    renderUpdateMenu(menuState);
-    renderRemoveMenu(menuState);
-    showCount(Object.keys(menuState).length);
-  }
-
-  function renderAppendMenu(menuState) {
-    for (const [seq, menu] of Object.entries(menuState)) {
-      const menuWrapperId = getMenuWrapperId(seq);
-
-      if (getById(menuWrapperId)) {
-        continue;
-      }
-
-      const { name, isSoldOut } = menu;
-
-      appendMenu(
-        seq,
-        name,
-        menuWrapperId,
-        {
-          onClickRemove: () => remove(seq),
-          onClickUpdate: () => update(seq),
-          onClickSoldOut: () => soldOut(seq),
-        },
-        isSoldOut
-      );
-    }
-  }
-
-  function renderSoldOutMenu(menuState) {
-    const { menuName } = elementIdMap;
-
-    for (const [seq, menu] of Object.entries(menuState)) {
-      const $menuName = getById(`${menuName}${seq}`);
-      const { isSoldOut } = menu;
-
-      if (isSoldOut && !$menuName.classList.contains("sold-out")) {
-        $menuName.classList.add("sold-out");
-      }
-
-      if (!isSoldOut && $menuName.classList.contains("sold-out")) {
-        $menuName.classList.remove("sold-out");
-      }
-    }
-  }
-
-  function renderUpdateMenu(menuState) {
-    const { menuName } = elementIdMap;
-
-    for (const [seq, menu] of Object.entries(menuState)) {
-      const { name } = menu;
-      const $menuName = getById(`${menuName}${seq}`);
-      if ($menuName.textContent !== name) {
-        $menuName.textContent = name;
-      }
-    }
-  }
-
-  function renderRemoveMenu(menuState) {
-    const { espressoMenuList } = elementIdMap;
-
-    for (const menu of getById(espressoMenuList).children) {
-      const seq = menu.id.replace("menuWrapper-", "");
-      if (!menuState[seq]) {
-        menu.remove();
-      }
-    }
-  }
-
-  function appendMenu(
-    seq,
-    name,
-    menuWrapperId,
-    { onClickSoldOut, onClickUpdate, onClickRemove },
-    isSoldOut
-  ) {
-    const { espressoMenuList: espressoMenuListId } = elementIdMap;
-
-    appendHtml(espressoMenuListId, $menuWrapper(menuWrapperId));
-    appendHtml(menuWrapperId, $menuName(seq, name, isSoldOut));
-    appendHtml(menuWrapperId, $soldOutButton(seq), {
-      eventName: "onclick",
-      callback: onClickSoldOut,
-    });
-    appendHtml(menuWrapperId, $updateButton(seq), {
-      eventName: "onclick",
-      callback: onClickUpdate,
-    });
-    appendHtml(menuWrapperId, $removeButton(seq), {
-      eventName: "onclick",
-      callback: onClickRemove,
-    });
-  }
-
-  function showCount(count) {
-    getById("count").textContent = count;
-  }
-
-  function $removeElement(seq) {
-    getById(getMenuWrapperId(seq)).remove();
-  }
-
-  function stashMenu() {
-    for (const [seq, _] of Object.entries(menuState)) {
-      $removeElement(seq);
-    }
-  }
-
-  /**
-   * Get html template functions
-   */
-
-  function $removeButton(seq) {
-    const { removeButton } = elementIdMap;
-    return `<button type="button" id="${removeButton}${seq}" class="bg-gray-50 text-gray-500 text-sm menu-remove-button">삭제</button>`;
-  }
-
-  function $updateButton(seq) {
-    const { updateButton } = elementIdMap;
-    return `<button type="button" id="${updateButton}${seq}" class="bg-gray-50 text-gray-500 text-sm mr-1 menu-edit-button">수정</button>`;
-  }
-
-  function $soldOutButton(seq) {
-    const { soldOutButton } = elementIdMap;
-    return `<button id="${soldOutButton}${seq}" type="button" class="bg-gray-50 text-gray-500 text-sm mr-1 menu-sold-out-button">
-          품절
-        </button>`;
-  }
-
-  function $menuName(seq, name, isSoldOut) {
-    const { menuName } = elementIdMap;
-    return `<span id="${menuName}${seq}" class="w-100 pl-2 menu-name ${
-      isSoldOut ? "sold-out" : ""
-    }">${name}</span>`;
-  }
-
-  function $menuWrapper(menuWrapperId) {
-    return `<li data-menu-id="${menuWrapperId}" id="${menuWrapperId}" class="menu-list-item d-flex items-center py-2"></li>`;
-  }
-
-  return [getMenuState, setMenuState, getMenuSeq, setCategoryName];
-}
-
 const onSubmit = (e) => {
-  const submitForm = e.target;
   e.preventDefault();
 
-  const name = getById("espresso-menu-name").value;
+  const { espressoMenuNameInput } = elementIdMap;
+  const name = getById(espressoMenuNameInput).value;
 
   if (!name) {
     return;
   }
+
+  const submitForm = e.target;
 
   setState({
     ...getState(),
@@ -365,10 +49,6 @@ function bindOnClickMenuCategory() {
   }
 }
 
-/**
- * When DOMContentLoaded
- */
-
 document.addEventListener(
   "DOMContentLoaded",
   () => {
@@ -377,21 +57,3 @@ document.addEventListener(
   },
   false
 );
-
-/**
- * Utils
- */
-
-function getById(id) {
-  return document.getElementById(id);
-}
-
-function appendHtml(parentId, htmlTemplate, event) {
-  const $ = getById(`${parentId}`);
-  $.insertAdjacentHTML("beforeend", htmlTemplate);
-
-  if (event) {
-    const { eventName, callback } = event;
-    $.lastChild[eventName] = callback;
-  }
-}
