+++ src/js/utils/constant_utils.js
@@ -1,0 +1,17 @@
+export const elementIdMap = {
+  espressoMenuForm: "espresso-menu-form",
+  espressoMenuList: "espresso-menu-list",
+  espressoMenuNameInput: "espresso-menu-name",
+  removeButton: "removeButton",
+  updateButton: "updateButton",
+  soldOutButton: "soldOutButton",
+  menuName: "menuName",
+  menuWrapper: "menuWrapper",
+  menuCategoryButtonWrapper: "menuCategoryButtonWrapper",
+  menuTitleName: "menuTitleName",
+};
+
+export const localStorageKey = {
+  categoryState: "moonbucksState.categoryState",
+  categorySeqState: "moonbucksState.categorySeqState",
+};
