+++ src/js/utils/control_dom_utils.js
@@ -1,0 +1,13 @@
+export function getById(id) {
+  return document.getElementById(id);
+}
+
+export function appendHtml(parentId, htmlTemplate, event) {
+  const $ = getById(`${parentId}`);
+  $.insertAdjacentHTML("beforeend", htmlTemplate);
+
+  if (event) {
+    const { eventName, callback } = event;
+    $.lastChild[eventName] = callback;
+  }
+}
